refactor(IdempotentAspect): 优化幂等切面逻辑并修改类访问权限
将IdempotentAspect类访问权限改为包私有，并增加未登录时的随机数处理逻辑
修复UserSavedWordServiceimpl中删除操作返回类型不匹配问题

+++ src/main/java/top/aprdec/onepractice/service/impl/UserSavedWordServiceimpl.java
@@ -175,7 +175,7 @@ public class UserSavedWordServiceimpl implements UserSavedWordService {
         }
     }
 
-
+    // 添加收藏关系（原子操作）
     // 查找单词（不创建新单词）
     private WordsDO findWord(String word) {
         // 优先查缓存
@@ -217,7 +217,7 @@ public class UserSavedWordServiceimpl implements UserSavedWordService {
             }
 
             // 从数据库中删除
-            int deletedRows = easyEntityQuery.deletable(UserSavedWordsDO.class)
+            Long deletedRows = easyEntityQuery.deletable(UserSavedWordsDO.class)
                     .where(u -> {
                         u.userId().eq(userId);
                         u.and(() -> {
@@ -235,7 +235,6 @@ public class UserSavedWordServiceimpl implements UserSavedWordService {
         }
     }
 
-    // 添加收藏关系（原子操作）
     private void addSavedWordMapping(Long userId, Long wordId) {
         String userSavedKey = RedisKeyConstant.USER_SAVED_WORD_LIST + userId;
